requirements.txt
numpy
Pillow
pytesseract
streamlit
python-dotenv
requests